park_ride/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow==10.3.0
psycopg2-binary==2.9.9   # For PostgreSQL (remove if using SQLite)
mysqlclient==2.2.4        # For MySQL (remove if not needed)
python-decouple==3.8      # For environment variable management
requests==2.32.2

